added exercise 06 - GCD

+++ HomeWork_01/04.PrintNumbers/04.PrintNumbers.cpp
@@ -9,7 +9,7 @@ int main()
 	int N;
 	cin >> N;
 
-	for (size_t i = 1; i <= N; i++)
+	for (int i = 1; i <= N; i++)
 	{
 		cout << i << " ";
 	}
